+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Bug fixes:
 
 Features:
 
+- Allow config files to include other config files via. `$include` and
+  `$include-glob` #989
 - Added `frame` function to create a new data frame within an expressoin.
 - Added `sum` and `count` functions #865
 - New component based report generator #851